Move storev2 watcher metrics to storev2 package
